Refactor test-component-model
